BB-190: Split and classify search components
